# swagger-spec.yaml
openapi: 3.1.0
paths:
  /authentication/login:
    get:
      operationId: authenticationLogin
      parameters:
        - name: code
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: ''
      tags: &a1
        - Authentication
  /authentication/me:
    get:
      operationId: authenticationGetMe
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticateUserOkDto'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponseDto'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponseDto'
      security:
        - defaultBearerAuth: []
      summary: ''
      tags: *a1
  /authentication/revoke-token:
    delete:
      operationId: authenticationRevokeToken
      parameters: []
      responses:
        '204':
          description: ''
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponseDto'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponseDto'
      security:
        - defaultBearerAuth: []
      summary: ''
      tags: *a1
  /tournament:
    get:
      operationId: tournamentGetAll
      parameters:
        - name: itemsPerPage
          required: false
          in: query
          schema:
            type: number
        - name: page
          required: false
          in: query
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentArrayOkResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentBadRequestResponseDto'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponseDto'
      summary: ''
      tags: &a2
        - Tournament
    post:
      operationId: tournamentCreate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TournamentBodyCreateDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentCreatedResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentBadRequestResponseDto'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponseDto'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponseDto'
      security:
        - defaultBearerAuth: []
      summary: ''
      tags: *a2
  /tournament/{id}:
    get:
      operationId: tournamentGetOneById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentOKResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentBadRequestResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentNotFoundResponseDto'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponseDto'
      summary: ''
      tags: *a2
    delete:
      operationId: tournamentRemoveById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentNotFoundResponseDto'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponseDto'
      security:
        - defaultBearerAuth: []
      summary: ''
      tags: *a2
info:
  title: Serezin-Gaming docs
  description: ''
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    defaultBearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    AuthenticateUserDto:
      type: object
      properties:
        discord:
          type: object
          properties:
            id:
              type: string
              required: true
            username:
              type: string
              required: true
            avatar:
              type: string
              required: true
            global_name:
              type: string
              nullable: true
        riot:
          type: object
          nullable: true
          properties:
            id:
              type: string
              required: true
            name:
              type: string
              required: true
      required:
        - discord
        - riot
    AuthenticateUserOkDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 200
        message:
          type: string
          example: Authenticate user fetch sucessfully
        data:
          example:
            - discord:
                id: '123456789012345678'
                username: test
                avatar: '12345678901234567890122345678912'
                global_name: Test
              riot:
                id: '123456789012345678901234567890123456789012345678901234567890123456789012345678'
                name: Test#EUW
          type: array
          items:
            $ref: '#/components/schemas/AuthenticateUserDto'
        meta:
          type: object
      required:
        - statusCode
        - message
        - data
        - meta
    ForbiddenResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 403
        error:
          type: object
          properties:
            message:
              oneOf:
                - type: string
                - type: array
                  items:
                    type: string
              example: Not authorized to access this resource
      required:
        - statusCode
        - error
    InternalServerErrorResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 500
        error:
          type: object
          properties:
            message:
              oneOf:
                - type: string
                - type: array
                  items:
                    type: string
              example: Internal server error
      required:
        - statusCode
        - error
    TournamentDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        owner:
          type: string
        members:
          type: array
          items:
            type: string
        isArchived:
          type: boolean
      required:
        - id
        - name
        - owner
        - members
        - isArchived
    ResponsePaginationMetaDto:
      type: object
      properties:
        currentPage:
          type: number
        totalPages:
          type: number
        itemsPerPage:
          type: number
        totalItems:
          type: number
        haveNextPage:
          type: boolean
        havePreviousPage:
          type: boolean
      required:
        - currentPage
        - totalPages
        - itemsPerPage
        - totalItems
        - haveNextPage
        - havePreviousPage
    TournamentArrayOkResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 200
        message:
          type: string
          example: Tournaments fetch sucessfully
        data:
          example:
            - &a3
              id: 51d0af06-f14b-4932-bba8-697201468cda
              name: Tournament Edition 2024
              owner: 60f51f4e-4e15-4c2b-9ba9-7675c850d81a
              members:
                - 60f51f4e-4e15-4c2b-9ba9-7675c850d81a
              isArchived: false
          type: array
          items:
            $ref: '#/components/schemas/TournamentDto'
        meta:
          $ref: '#/components/schemas/ResponsePaginationMetaDto'
      required:
        - statusCode
        - message
        - data
        - meta
    TournamentBadRequestResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        error:
          type: object
          properties:
            message:
              oneOf:
                - type: string
                - type: array
                  items:
                    type: string
              example: Tournament data malformed
      required:
        - statusCode
        - error
    TournamentOKResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 200
        message:
          type: string
          example: Tournament fetch sucessfully
        data:
          example: *a3
          allOf:
            - $ref: '#/components/schemas/TournamentDto'
        meta:
          type: object
      required:
        - statusCode
        - message
        - data
        - meta
    TournamentNotFoundResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 404
        error:
          type: object
          properties:
            message:
              oneOf:
                - type: string
                - type: array
                  items:
                    type: string
              example: Tournament not found
      required:
        - statusCode
        - error
    TournamentBodyCreateDto:
      type: object
      properties:
        name:
          type: string
          default: Tournament Edition 2024
      required:
        - name
    TournamentCreatedResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 201
        message:
          type: string
          example: Tournament create sucessfully
        data:
          example: *a3
          allOf:
            - $ref: '#/components/schemas/TournamentDto'
        meta:
          type: object
      required:
        - statusCode
        - message
        - data
        - meta